added retrofit builder, api service and DTO for feature:page_two, renamed modules
