+++ .idea/gradle.xml
@@ -14,11 +14,11 @@
             <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/core/dagger" />
+            <option value="$PROJECT_DIR$/core/detailed_info_api" />
+            <option value="$PROJECT_DIR$/core/detailed_info_api_internal" />
             <option value="$PROJECT_DIR$/core/latest_api" />
             <option value="$PROJECT_DIR$/core/latest_api_internal" />
             <option value="$PROJECT_DIR$/core/navigation" />
-            <option value="$PROJECT_DIR$/core/page_two_api" />
-            <option value="$PROJECT_DIR$/core/page_two_api_internal" />
             <option value="$PROJECT_DIR$/core/sale_api" />
             <option value="$PROJECT_DIR$/core/sale_api_internal" />
             <option value="$PROJECT_DIR$/core/ui" />
