+++ core/detailed_info_api/.gitignore
