+++ core/detailed_info_api/proguard-rules.pro
