+++ core/detailed_info_api/src/main/AndroidManifest.xml
