+++ core/detailed_info_api_internal/.gitignore
