+++ core/detailed_info_api_internal/proguard-rules.pro
