+++ core/detailed_info_api_internal/src/main/AndroidManifest.xml
