+++ settings.gradle
@@ -30,5 +30,5 @@ include ':utils:network'
 include ':storage:user_database'
 include ':storage:user_database_api'
 include ':utils:livedata'
-include ':core:page_two_api_internal'
-include ':core:page_two_api'
+include ':core:detailed_info_api_internal'
+include ':core:detailed_info_api'
